digest: cleanup digester and verifier creation
