+++ digest/verifiers.go
@@ -17,8 +17,7 @@ type Verifier interface {
 	Verified() bool
 }
 
-// NewDigestVerifier returns a verifier that compares the written bytes
-// against a passed in digest.
+// NewDigestVerifier is deprecated. Please use Digest.Verifier.
 func NewDigestVerifier(d Digest) (Verifier, error) {
 	if err := d.Validate(); err != nil {
 		return nil, err
